DB-3856: Compaction-based checkpointing.
When performing a compaction, we already perform transactional
resolution to remove rolled back versions and add commit timestamps to
data. This extends that previous behavior to also perform checkpointing
operations. In particular, during major compactions, this

1. Merges together all updates below the MinimumActiveTransaction(MAT)
2. removes all cells below that MAT
3. Removes all rows which have been tombstoned below the MAT

Technically speaking, condition #3 introduces a "phantom read" scenario,
when a reader first reads a row as being present, then another
transaction deletes the row and commits. Upon re-reading that row, the
first reader does not see the row any more. This is a phantom read,
which is allowed by Snapshot Isolation (since we are technically only at
the REPEATED_READS ANSI isolation level), but we'll need to resolve this
for SERIALIZABLE transactions.

In order to do this, we must alse pause transactional maintenance operators during
compaction.

When compaction is ongoing, then we pause ReadResolution. Otherwise,
there is a race condition where a ReadResolver can add a commit
timestamp to a row that has been deleted (because it was tombstoned
below the MAT). Pausing ReadResolution ensures that this race condition
is not hit.

The same logic applies to checkpoint resolution and RollForwards.

+++ splice_constants/src/main/java/com/splicemachine/constants/SIConstants.java
@@ -160,7 +160,7 @@ public class SIConstants extends SpliceConstants {
      */
     @Parameter public static final String CHECKPOINT_SEEK_THRESHOLD = "splice.txn.checkpoint.seekThreshold";
     @DefaultValue(CHECKPOINT_SEEK_THRESHOLD) public static final int DEFAULT_CHECKPOINT_SEEK_THRESHOLD=SpliceConstants.scannerBatchSize/2;
-    public static long checkpointSeekThreshold;
+    public static int checkpointSeekThreshold;
 
     public static long matWatcherRefreshMs;
     public static void setParameters(Configuration config){
